app/requirements.txt
torch
torchvision
Pillow
transformers
flask
google-generativeai
werkzeug
